requirements.txt
# Core PyTorch
torch>=1.9.0
torchvision>=0.10.0

# Segmentation Models
segmentation-models-pytorch>=0.3.0

# Data Processing
albumentations>=1.3.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
einops>=0.6.0
scikit-image>=0.19.0

# Optional: For advanced features
tensorboard>=2.8.0
wandb>=0.13.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development (optional)
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
